tests/wpt/web-platform-tests/html/browsers/browsing-the-web/overlapping-navigations-and-traversals/same-document-nav-cross-document-nav.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Cross-document navigation after a same-document navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<!--
  These tests are kind of silly since it's hard to imagine any other result:
  same-document navigations are always synchronous so of course the
  same-document navigation will succeed, followed by the cross-document one.

  Nevertheless they're nice as a basis from which to write corresponding app
  history tests, where the consequences aren't as obvious.
-->

<body>
<script type="module">
import { createIframe, waitForLoad } from "./resources/helpers.mjs";

promise_test(async t => {
  const iframe = await createIframe(t);

  iframe.contentWindow.location.hash = "#1";
  assert_equals(iframe.contentWindow.location.hash, "#1");

  iframe.contentWindow.location.search = "?2";
  await waitForLoad(iframe);
  assert_equals(iframe.contentWindow.location.search, "?2");
}, "fragment navigation then cross-document navigation");

promise_test(async t => {
  const iframe = await createIframe(t);

  iframe.contentWindow.history.pushState(null, "", "?1");
  assert_equals(iframe.contentWindow.location.search, "?1");

  iframe.contentWindow.location.search = "?2";
  await waitForLoad(iframe);
  assert_equals(iframe.contentWindow.location.search, "?2");
}, "pushState() then cross-document navigation");
</script>
